move IdentityProvider and User interfaces to app/types
