+++ server/sonar-web/src/main/js/app/types.ts
@@ -254,3 +254,25 @@ export enum RuleInheritance {
   Inherited = 'INHERITED',
   Overridden = 'OVERRIDES'
 }
+
+export interface IdentityProvider {
+  backgroundColor: string;
+  helpMessage?: string;
+  iconPath: string;
+  key: string;
+  name: string;
+}
+
+export interface User {
+  active: boolean;
+  avatar?: string;
+  email?: string;
+  externalIdentity?: string;
+  externalProvider?: string;
+  groups?: string[];
+  local: boolean;
+  login: string;
+  name: string;
+  scmAccounts?: string[];
+  tokensCount?: number;
+}
