+++ server/sonar-web/src/main/js/apps/users/UsersApp.tsx
@@ -25,8 +25,8 @@ import Header from './Header';
 import ListFooter from '../../components/controls/ListFooter';
 import Search from './Search';
 import UsersList from './UsersList';
-import { getIdentityProviders, IdentityProvider, searchUsers, User } from '../../api/users';
-import { Paging } from '../../app/types';
+import { getIdentityProviders, searchUsers } from '../../api/users';
+import { Paging, IdentityProvider, User } from '../../app/types';
 import { translate } from '../../helpers/l10n';
 import { parseQuery, Query, serializeQuery } from './utils';
