+++ server/sonar-web/src/main/js/apps/users/UsersList.tsx
@@ -19,7 +19,7 @@
  */
 import * as React from 'react';
 import UserListItem from './components/UserListItem';
-import { IdentityProvider, User } from '../../api/users';
+import { IdentityProvider, User } from '../../app/types';
 import { translate } from '../../helpers/l10n';
 
 interface Props {
